[Prover Service] Reshuffle common test code.
