+++ prover-service/src/tests/training_wheels.rs
@@ -1,8 +1,9 @@
 // Copyright (c) Aptos Foundation
 
 use crate::request_handler::training_wheels;
-use crate::tests::common::types::{ProofTestCase, TestJWTPayload};
-use crate::tests::common::{gen_test_jwk_keypair, types::TestJWKKeyPair};
+use crate::tests::types::TestJWKKeyPair;
+use crate::tests::types::{ProofTestCase, TestJWTPayload};
+use crate::tests::utils;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 #[test]
@@ -25,11 +26,11 @@ fn test_validate_jwt_invalid_signature() {
 
     // Create a test case and convert it to a prover request input
     let testcase = ProofTestCase::default_with_payload(jwt_payload).compute_nonce();
-    let jwk_keypair = gen_test_jwk_keypair();
+    let jwk_keypair = utils::gen_test_jwk_keypair();
     let prover_request_input = testcase.convert_to_prover_request(&jwk_keypair);
 
     // Verify the JWT signature using a different keypair to simulate an invalid signature
-    let another_jwk_keypair = gen_test_jwk_keypair();
+    let another_jwk_keypair = utils::gen_test_jwk_keypair();
     let result = training_wheels::validate_jwt_signature(
         &another_jwk_keypair.into_rsa_jwk(),
         &prover_request_input.jwt_b64,
@@ -58,7 +59,7 @@ fn test_validate_jwt_sig_and_dates_expired() {
 fn test_jwt_signature_validation(jwt_payload: TestJWTPayload, expect_success: bool) {
     // Create a test case and convert it to a prover request input
     let testcase = ProofTestCase::default_with_payload(jwt_payload).compute_nonce();
-    let jwk_keypair = gen_test_jwk_keypair();
+    let jwk_keypair = utils::gen_test_jwk_keypair();
     let prover_request_input = testcase.convert_to_prover_request(&jwk_keypair);
 
     // Verify the JWT signature
