+++ prover-service/src/tests/utils.rs
@@ -1,12 +1,13 @@
 // Copyright (c) Aptos Foundation
 
-use anyhow::Result;
+use crate::tests::types::{DefaultTestJWKKeyPair, TestJWKKeyPair};
 use num_bigint::BigUint;
-use rsa::{
-    pkcs1::{EncodeRsaPrivateKey, LineEnding},
-    traits::PublicKeyParts,
-};
+use rsa::pkcs1::{EncodeRsaPrivateKey, LineEnding};
+use rsa::rand_core;
+use rsa::traits::PublicKeyParts;
 use std::str::FromStr;
+
+const TEST_JWK_EXPONENT_STR: &str = "65537";
 
 #[derive(Debug, PartialEq, Eq)]
 pub struct RsaPublicKey {
@@ -15,7 +16,7 @@ pub struct RsaPublicKey {
 
 impl RsaPublicKey {
     #[allow(clippy::all)]
-    pub fn from_str(s: &str) -> Result<Self, anyhow::Error> {
+    pub fn from_str(s: &str) -> anyhow::Result<Self, anyhow::Error> {
         Ok(Self {
             modulus: BigUint::from_str(s)?,
         })
@@ -36,7 +37,7 @@ impl RsaPublicKey {
     }
 
     // TODO test from and as below
-    pub fn from_mod_b64(modulus_b64: &str) -> Result<Self, anyhow::Error> {
+    pub fn from_mod_b64(modulus_b64: &str) -> anyhow::Result<Self, anyhow::Error> {
         let modulus_bytes = base64::decode_config(modulus_b64, base64::URL_SAFE_NO_PAD)?;
         Ok(RsaPublicKey::from_bytes(&modulus_bytes))
     }
@@ -55,7 +56,7 @@ impl RsaPrivateKey {
         rng: &mut R,
         bit_size: usize,
         exp: &BigUint,
-    ) -> Result<Self, anyhow::Error>
+    ) -> anyhow::Result<Self, anyhow::Error>
     where
         R: rsa::rand_core::CryptoRngCore + ?Sized,
     {
@@ -81,9 +82,21 @@ impl RsaPrivateKey {
 impl From<&RsaPrivateKey> for RsaPublicKey {
     fn from(value: &RsaPrivateKey) -> Self {
         RsaPublicKey {
-            modulus: num_bigint::BigUint::from_bytes_be(
-                &value.internal_private_key.n().to_bytes_be(),
-            ),
+            modulus: BigUint::from_bytes_be(&value.internal_private_key.n().to_bytes_be()),
         }
     }
+}
+
+pub fn gen_test_jwk_keypair() -> impl TestJWKKeyPair {
+    gen_test_jwk_keypair_with_kid_override("test-rsa")
+}
+
+pub fn gen_test_jwk_keypair_with_kid_override(kid: &str) -> impl TestJWKKeyPair {
+    let mut rng = rand_core::OsRng;
+    DefaultTestJWKKeyPair::new_with_kid_and_exp(
+        &mut rng,
+        kid,
+        num_bigint::BigUint::from_str(TEST_JWK_EXPONENT_STR).unwrap(),
+    )
+    .unwrap()
 }
